Move models reader to cli namespace
