+++ GPTifier/include/cli.hpp
@@ -1,0 +1,6 @@
+#pragma once
+
+namespace cli
+{
+bool get_opts_models(const int argc, char **argv);
+}
