+++ GPTifier/src/command_models.cpp
@@ -1,49 +1,18 @@
 #include "command_models.hpp"
 
 #include "api.hpp"
+#include "cli.hpp"
 #include "help_messages.hpp"
 #include "reporting.hpp"
 #include "utils.hpp"
 
 #include <fmt/core.h>
-#include <getopt.h>
 #include <iostream>
 #include <json.hpp>
-#include <stdexcept>
 #include <string>
 
 namespace
 {
-
-bool read_cli_models(const int argc, char **argv)
-{
-    bool print_help = false;
-
-    while (true)
-    {
-        static struct option long_options[] = {{"help", no_argument, 0, 'h'}, {0, 0, 0, 0}};
-
-        int option_index = 0;
-        int c = getopt_long(argc, argv, "h", long_options, &option_index);
-
-        if (c == -1)
-        {
-            break;
-        }
-
-        switch (c)
-        {
-        case 'h':
-            print_help = true;
-            break;
-        default:
-            std::cerr << "Try running with -h or --help for more information\n";
-            exit(EXIT_FAILURE);
-        }
-    }
-
-    return print_help;
-}
 
 void print_row(const std::string &id, const std::string &owned_by, const std::string &creation_time)
 {
@@ -80,7 +49,7 @@ void print_models_response(const std::string &response)
 
 void command_models(const int argc, char **argv)
 {
-    bool print_help = read_cli_models(argc, argv);
+    bool print_help = cli::get_opts_models(argc, argv);
 
     if (print_help)
     {
